menu

+++ backend/src/main/java/com/portfolio/backend/menu/repository/MenuRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface MenuRepository extends JpaRepository<Menu, Long> {
+public interface MenuRepository extends JpaRepository<Menu, Integer> {
 
     List<Menu> findAllByOrderBySortOrderAsc();
 }
